Refactor centromeres code into separate module
