+++ scripts/python/fetch_chromosomes/fetch_cytobands_from_dbs.py
@@ -56,6 +56,7 @@ from .utils import *
 from .ucsc import *
 from .ensembl import *
 from .genomaize import *
+from .centromeres import *
 
 if os.path.exists(output_dir) is False:
     os.mkdir(output_dir)
@@ -94,130 +95,6 @@ time_ucsc = 0
 time_ensembl = 0
 
 times = {'ncbi': 0, 'ucsc': 0, 'ensembl': 0}
-
-
-def merge_centromeres(bands_by_chr, centromeres):
-    """Adds p and q arms to cytobands; thus adds centromere to each chromosome.
-
-    This is a special case for Zea mays (maize, i.e. corn).
-    Ensembl Genomes provides band data with no cytogenetic arm assignment.
-    Genomaize provides centromere positions for each chromosome.
-    This function merges those two datasets to provide input directly
-    useable to Ideogram.js.
-    """
-    logger.info('Entering merge_centromeres')
-    new_bands = {}
-
-    for chr in bands_by_chr:
-        bands = bands_by_chr[chr]
-        new_bands[chr] = []
-        centromere = centromeres[chr]
-        cen_start, cen_stop = centromere
-        pcen_index = None
-
-        j = 0
-        for i, band in enumerate(bands):
-            new_band = band
-            band_start, band_stop = band[1:3]
-            if int(band_stop) < int(cen_start):
-                arm = 'p'
-            else:
-                arm = 'q'
-
-                if int(band_start) < int(cen_stop):
-                    # Omit any q-arm bands that start before q-arm pericentromeric band
-                    if chr == '1':
-                        logger.info('Omit band:')
-                        logger.info(band)
-                    j += 1
-                    continue
-
-                if pcen_index is None:
-                    pcen_index = i - j
-
-                    # Extend nearest p-arm band's stop coordinate to the
-                    # p_cen's start coordinate (minus 1)
-                    cen_start_pre = str(int(cen_start) - 1)
-                    new_bands[chr][i - j - 1][3] = cen_start_pre
-                    new_bands[chr][i - j - 1][5  ] = cen_start_pre
-
-                    # Extend nearest q-arm band's start coordinate to the
-                    # q_cen's stop coordinate (plus 1)
-                    cen_stop_post = str(int(cen_stop) + 1)
-                    bands[i + j][1] = cen_stop_post
-                    bands[i + j][3] = cen_stop_post
-
-                    # Coordinates of the centromere itself
-                    cen_mid = int(cen_start) + round((int(cen_stop)-int(cen_start))/2)
-
-                    pcen = [
-                        'p', 'pcen', cen_start, str(cen_mid - 1),
-                        cen_start, str(cen_mid - 1), 'acen'
-                    ]
-                    qcen = [
-                        'q', 'qcen', str(cen_mid), cen_stop,
-                        str(cen_mid), cen_stop, 'acen'
-                    ]
-            new_band.insert(0, arm)
-            new_bands[chr].append(new_band)
-        if pcen_index is not None:
-            new_bands[chr].insert(pcen_index, qcen)
-            new_bands[chr].insert(pcen_index, pcen)
-    return new_bands
-
-
-def parse_centromeres(bands_by_chr):
-    """Adds p and q arms to cytobands, by parsing embedded centromere bands.
-
-    This is a special case for assigning cytogenetic arms to certain organisms
-    from Ensembl Genomes, including: Aspergillus fumigatus, Aspergillus
-    nidulans, Aspergillus niger, Aspergillus oryzae (various fungi);
-    Oryza sativa (rice); and Hordeum vulgare (barley).
-
-    Bands are assigned an arm based on their position relative to the embedded
-    centromere.
-    """
-    logger.info('Entering parse_centromeres')
-
-    # If centromeres aren't embedded in the input banding data,
-    # then simply return the input without modification.
-    has_centromere = False
-    for chr in bands_by_chr:
-        bands = bands_by_chr[chr]
-        for band in bands:
-            stain = band[-1]
-            if stain == 'acen':
-                has_centromere = True
-    if has_centromere is False:
-        return bands_by_chr
-
-    new_bands = {}
-
-    for chr in bands_by_chr:
-        bands = bands_by_chr[chr]
-        new_bands[chr] = []
-
-        # On each side of the centromere -- the p-arm side and the q-arm
-        # side -- there is a band with a "stain" value of "acen".  Here,
-        # we find the index of the acen band on the p-arm side.  That
-        # band and all bands to the left of it are on the p arm.  All
-        # bands to the right of it are on the q arm.
-        pcen_index = None
-        for i, band in enumerate(bands):
-            stain = band[-1]
-            if stain == 'acen':
-                pcen_index = i
-        for i, band in enumerate(bands):
-            arm = ''
-            if pcen_index is not None:
-                if i < pcen_index:
-                    arm = 'p'
-                else:
-                    arm = 'q'
-            band.insert(0, arm)
-            new_bands[chr].append(band)
-
-    return new_bands
 
 
 def patch_telomeres(bands_by_chr):
